Create admin role for user to manage other users

+++ src/todo/todo.controller.ts
@@ -10,6 +10,7 @@ import {
   Req,
   HttpCode,
   HttpStatus,
+  Session,
 } from '@nestjs/common';
 import { TodoService } from './todo.service';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
@@ -22,23 +23,44 @@ export class TodoController {
 
   @Post()
   @HttpCode(HttpStatus.CREATED)
-  async create(@Req() req, @Body() createTodoDto: CreateTodoDto) {
-    const currentUser = req.user;
+  async create(
+    @Req() req,
+    @Body() createTodoDto: CreateTodoDto,
+    @Session() session: Record<string, any>,
+  ) {
+    console.log(session.user);
+
+    const currentUser = session.user;
+
+    if (currentUser.isBlocked) {
+      return { message: 'You are blocked. Cannot create a todo.' };
+    }
 
     const todo = await this.todoService.create(createTodoDto, currentUser);
     return todo;
   }
 
   @Get()
-  async findAll(@Req() req) {
-    const userId = req.user.id;
+  async findAll(@Req() req, @Session() session: Record<string, any>) {
+    const userId = session.user.id;
+
+    if (session.user.isBlocked) {
+      return { message: 'You are blocked. Cannot create a todo.' };
+    }
+
     const todos = await this.todoService.findAll(userId);
+
     return todos;
   }
 
   @Get(':id')
   async findOne(@Req() req, @Param('id') id: number) {
     const userId = req.user.id;
+
+    if (req.user.isBlocked) {
+      return { message: 'You are blocked. Cannot access this todo.' };
+    }
+
     const todo = await this.todoService.findOne(id, userId);
     if (!todo) {
       return { message: 'Todo not found' };
@@ -49,6 +71,11 @@ export class TodoController {
   @Put(':id')
   async update(@Req() req, @Param('id') id: number, @Body() body) {
     const userId = req.user.id;
+
+    if (req.user.isBlocked) {
+      return { message: 'You are blocked. Cannot access this todo.' };
+    }
+
     const todo = await this.todoService.update(id, userId, body);
     if (!todo) {
       return { message: 'Todo not found' };
@@ -59,6 +86,11 @@ export class TodoController {
   @Delete(':id')
   async remove(@Req() req, @Param('id') id: number) {
     const userId = req.user.id;
+
+    if (req.user.isBlocked) {
+      return { message: 'You are blocked. Cannot access this todo.' };
+    }
+
     await this.todoService.remove(id, userId);
     return { message: 'Todo deleted' };
   }
